request_test/req_test/api.py
import json
import requests
from . import conf as cnf
from rest_framework.views import APIView
from rest_framework.response import Response
from rest_framework.exceptions import ParseError, ValidationError
