request_test/requirement.txt
Django==2.2.7
djangorestframework==3.10.3
requests==2.22.0
django-cors-headers==3.2.0
django-rest-swagger==2.2.0